+++ CHANGELOG.md
@@ -6,14 +6,15 @@
 
 - **#794** When a keycap on the virtual keyboard with associated alternate keys
   is long pressed, the other UI elements on the screen are ignored (a scrim
-  element is inserted behind the panel to capture events). This also adresses an
-  issue on iPad OS where the document selection could be altered after
+  element is inserted behind the panel to capture events).
+- On iPad OS prevent the document selection from being altered after
   long-pressing an alternate key in some cases.
 
 ### Bug Fixes
 
 - A `\chi13` (0.1em) gap between the nucleus and the above element was missing
   in `OverUnder` atoms (`\overset`, etc...).
+- On Safari iOS, correctly display the keyboard toggle glyph.
 
 ## 0.61.0 (2021-04-22)